update: fetchCompletedQuestions está agora em utils para ser usado em Home
